requirements.txt
arrow==1.2.2
asgiref==3.5.0
binaryornot==0.4.4
black==22.3.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
cookiecutter==1.7.3
Django==4.0.1
django-browser-reload==1.3.0
django-filter==21.1
django-tailwind==3.1.1
djangorestframework==3.13.1
fontawesomefree==6.0.0
idna==3.3
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
poyo==0.5.0
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
